change: Separated component class files.
All component classes now are located in their own files for easier use.
They can still be used from the components root class, they will be lazy
required when indexed.

Added luals annotations for everything, so it's really nice to use the
plugin now.

README was also updated to reflect these changes, and it was made a lot
shorter. It no longer goes into as much depth about properties and other
details, but that information is in the annotations.