refactor!: removed metatables and component setups
Extensive use of index metatables made the whole system very convoluted
and hard to work with, even for the person who made the system. That's a
bad sign. The solution was pretty much to completely remove all the
metatable nonsense so now you have to set up each component manually. If
you want a global config you can just make a table yourself and use it
in multiple places.